REQ sketch SortedView now uses the same SV as KllFloats.
