+++ src/test/java/org/apache/datasketches/quantilescommon/CrossCheckQuantilesTest.java
@@ -29,9 +29,7 @@ import static org.apache.datasketches.quantilescommon.LinearRanksAndQuantiles.ge
 import static org.apache.datasketches.quantilescommon.LinearRanksAndQuantiles.getTrueItemRank;
 import static org.apache.datasketches.quantilescommon.QuantileSearchCriteria.EXCLUSIVE;
 import static org.apache.datasketches.quantilescommon.QuantileSearchCriteria.INCLUSIVE;
-import static org.apache.datasketches.quantilescommon.ReflectUtilityTest.DOUBLES_SV_CTOR;
-import static org.apache.datasketches.quantilescommon.ReflectUtilityTest.KLL_FLOATS_SV_CTOR;
-import static org.apache.datasketches.quantilescommon.ReflectUtilityTest.REQ_SV_CTOR;
+
 import static org.testng.Assert.assertEquals;
 
 import java.util.Comparator;
@@ -46,7 +44,6 @@ import org.apache.datasketches.quantiles.DoublesSketch;
 import org.apache.datasketches.quantiles.ItemsSketch;
 import org.apache.datasketches.quantiles.UpdateDoublesSketch;
 import org.apache.datasketches.req.ReqSketch;
-import org.apache.datasketches.req.ReqSketchSortedView;
 import org.testng.annotations.Test;
 
 /**
@@ -140,12 +137,10 @@ public class CrossCheckQuantilesTest {
   KllItemsSketch<String> kllItemsSk = null;
   ItemsSketch<String> itemsSk = null;
 
-  ReqSketchSortedView reqFloatsSV = null;
-  FloatsSketchSortedView kllFloatsSV = null;
-  DoublesSketchSortedView kllDoublesSV = null;
-  DoublesSketchSortedView classicDoublesSV = null;
-  ItemsSketchSortedView<String> kllItemsSV = null;
+  FloatsSketchSortedView floatsSV = null;
+  DoublesSketchSortedView doublesSV = null;
   ItemsSketchSortedView<String> classicItemsSV = null;
+  ItemsSketchSortedView<String> kllItemsSV = null;
 
   public CrossCheckQuantilesTest() {}
 
@@ -174,11 +169,10 @@ public class CrossCheckQuantilesTest {
     float maxFloatvalue = getMaxFloatValue(set);
     for (float v = 5f; v <= maxFloatvalue + 5f; v += 5f) {
       trueRank = getTrueFloatRank(svCumWeights[set], svFValues[set],v, crit);
-      testRank = reqFloatsSV.getRank(v, crit);
+
+      testRank = floatsSV.getRank(v, crit);
       assertEquals(testRank, trueRank);
       testRank = reqFloatsSk.getRank(v, crit);
-      assertEquals(testRank, trueRank);
-      testRank = kllFloatsSV.getRank(v, crit);
       assertEquals(testRank, trueRank);
       testRank = kllFloatsSk.getRank(v, crit);
       assertEquals(testRank, trueRank);
@@ -191,12 +185,9 @@ public class CrossCheckQuantilesTest {
     for (double v = 5; v <= maxDoubleValue + 5; v += 5) {
       trueRank = getTrueDoubleRank(svCumWeights[set], svDValues[set],v, crit);
 
-      testRank = kllDoublesSV.getRank(v, crit);
+      testRank = doublesSV.getRank(v, crit);
       assertEquals(testRank, trueRank);
       testRank = kllDoublesSk.getRank(v, crit);
-      assertEquals(testRank, trueRank);
-
-      testRank = classicDoublesSV.getRank(v, crit);
       assertEquals(testRank, trueRank);
       testRank = classicDoublesSk.getRank(v, crit);
       assertEquals(testRank, trueRank);
@@ -216,7 +207,6 @@ public class CrossCheckQuantilesTest {
       assertEquals(testRank, trueRank);
       testRank = kllItemsSk.getRank(s, crit);
       assertEquals(testRank, trueRank);
-
       testRank = classicItemsSV.getRank(s, crit);
       assertEquals(testRank, trueRank);
       testRank = itemsSk.getRank(s, crit);
@@ -237,12 +227,9 @@ public class CrossCheckQuantilesTest {
       double normRank = i / dTwoN;
       trueFQ = getTrueFloatQuantile(svCumWeights[set], svFValues[set], normRank, crit);
 
-      testFQ = reqFloatsSV.getQuantile(normRank, crit);
+      testFQ = floatsSV.getQuantile(normRank, crit);
       assertEquals(testFQ, trueFQ);
       testFQ = reqFloatsSk.getQuantile(normRank, crit);
-      assertEquals(testFQ, trueFQ);
-
-      testFQ = kllFloatsSV.getQuantile(normRank, crit);
       assertEquals(testFQ, trueFQ);
       testFQ = kllFloatsSk.getQuantile(normRank, crit);
       assertEquals(testFQ, trueFQ);
@@ -257,12 +244,9 @@ public class CrossCheckQuantilesTest {
       double normRank = i / dTwoN;
       trueDQ = getTrueDoubleQuantile(svCumWeights[set], svDValues[set], normRank, crit);
 
-      testDQ = kllDoublesSV.getQuantile(normRank, crit);
+      testDQ = doublesSV.getQuantile(normRank, crit);
       assertEquals(testDQ, trueDQ);
       testDQ = kllDoublesSk.getQuantile(normRank, crit);
-      assertEquals(testDQ, trueDQ);
-
-      testDQ = classicDoublesSV.getQuantile(normRank, crit);
       assertEquals(testDQ, trueDQ);
       testDQ = classicDoublesSk.getQuantile(normRank, crit);
       assertEquals(testDQ, trueDQ);
@@ -281,7 +265,6 @@ public class CrossCheckQuantilesTest {
       assertEquals(testIQ, trueIQ);
       testIQ = kllItemsSk.getQuantile(normRank, crit);
       assertEquals(testIQ, trueIQ);
-
       testIQ = classicItemsSV.getQuantile(normRank, crit);
       assertEquals(testIQ, trueIQ);
       testIQ = itemsSk.getQuantile(normRank, crit);
@@ -289,7 +272,6 @@ public class CrossCheckQuantilesTest {
 
       println("Items set: " + set + ", rank: " + normRank + ", Q: " + trueIQ + ", crit: " + crit.toString());
     }
-
   }
 
   private double getMaxDoubleValue(int set) {
@@ -331,40 +313,19 @@ public class CrossCheckQuantilesTest {
   /*******BUILD & LOAD SVs***********/
 
   private void buildSVs(int set) throws Exception {
-    reqFloatsSV = getRawReqSV(svFValues[set], svCumWeights[set], totalN[set],
+    floatsSV = new FloatsSketchSortedView(svFValues[set], svCumWeights[set], totalN[set],
         svMaxFValues[set], svMinFValues[set]);
-    kllFloatsSV = getRawKllFloatsSV(svFValues[set], svCumWeights[set], totalN[set],
-        svMaxFValues[set], svMinFValues[set]);
-    kllDoublesSV = getRawDoublesSV(svDValues[set], svCumWeights[set], totalN[set],
-        svMaxDValues[set], svMinDValues[set]);
-    classicDoublesSV = getRawDoublesSV(svDValues[set], svCumWeights[set], totalN[set],
+    doublesSV = new DoublesSketchSortedView(svDValues[set], svCumWeights[set], totalN[set],
         svMaxDValues[set], svMinDValues[set]);
     String svImax = svIValues[set][svIValues[set].length - 1];
     String svImin = svIValues[set][0];
-    double normRankErr = KllSketch.getNormalizedRankError(k, true);
+
     kllItemsSV = new ItemsSketchSortedView<>(svIValues[set], svCumWeights[set], totalN[set],
-        comparator, svImax, svImin, normRankErr);
-    normRankErr = ItemsSketch.getNormalizedRankError(k, true);
+        comparator, svImax, svImin, KllSketch.getNormalizedRankError(k, true));
+
     classicItemsSV = new ItemsSketchSortedView<>(svIValues[set], svCumWeights[set], totalN[set],
-        comparator, svImax, svImin, normRankErr);
-  }
+        comparator, svImax, svImin, ItemsSketch.getNormalizedRankError(k, true));
 
-  private final static ReqSketchSortedView getRawReqSV(
-      final float[] values, final long[] cumWeights, final long totalN, final float maxItem, final float minItem)
-          throws Exception {
-    return (ReqSketchSortedView) REQ_SV_CTOR.newInstance(values, cumWeights, totalN, maxItem, minItem);
-  }
-
-  private final static FloatsSketchSortedView getRawKllFloatsSV(
-      final float[] values, final long[] cumWeights, final long totalN, final float maxItem, final float minItem)
-          throws Exception {
-    return (FloatsSketchSortedView) KLL_FLOATS_SV_CTOR.newInstance(values, cumWeights, totalN, maxItem, minItem);
-  }
-
-  private final static DoublesSketchSortedView getRawDoublesSV(
-      final double[] values, final long[] cumWeights, final long totalN, final double maxItem, final double minItem)
-          throws Exception {
-    return (DoublesSketchSortedView) DOUBLES_SV_CTOR.newInstance(values, cumWeights, totalN, maxItem, minItem);
   }
 
   /********BUILD DATA SETS**********/
